Moved all the chunks into a single buffer before exporting. This is slow and take up tons of memory. Don't use this revision unless you are willing to wait and have the program crash every once in a while.

+++ trunk/src/org/jmc/Chunk.java
@@ -122,13 +122,13 @@ public class Chunk {
 		 */
 		public Blocks(int num)
 		{
-			id=new int[num];
+			id=new short[num];
 			data=new byte[num];
 		}
 		/**
 		 * Block IDs.
 		 */
-		public int [] id;
+		public short [] id;
 		/**
 		 * Block meta-data.
 		 */
